Json and Gson Implemantation
This project was to show how to read from remote server api.The out put
is in JSON string format . i used Gson as this has proved to be a faster
way of parsing json in android java this also uses a recyclerview to
display the data.

also consider that this comes with several concepts in android on
network requests.

Feel free to fork and download the project for learning reasons.